feat: update scroll nav & scrollspy in vanilla
