+++ Vanilla/app/srcs/components/index.ts
@@ -18,3 +18,4 @@ export { default as onoffswitch } from './onoffswitch/onoffswitch';
 export { default as dropdown } from './dropdown/dropdown';
 export { default as parallexscroll } from './parallexscroll/parallexscroll';
 export { default as pointline } from './pointline/pointline';
+export { default as scrollnav } from './scrollnav/scrllnav';
